requirements.txt
flask
opencv-python
numpy
pillow
ultralytics
torch
albumentations
pyyaml
tqdm
# Optional for TFLite testing
# tflite-runtime
